// src/webapp/classicladder/src/generated/classicladder_client.ts
export enum ModbusReqType {
  COIL_READ = 0,
  COIL_WRITE = 1,
  INPUT_READ = 2,
  HOLDING_READ = 3,
  REGISTER_WRITE = 4,
  DIAGNOSTIC = 5,
}
// ...
export interface ModbusComParams {
  serialPort: string;
  serialSpeed: number;
  serialDataBits: number;
  serialStopBits: number;
  serialParity: number;
  serialUseRts: boolean;
  elementOffset: number;
  timeInterFrame: number;
  timeOutReceipt: number;
  timeAfterTransmit: number;
  debugLevel: number;
  mapCoilRead: number;
  mapCoilWrite: number;
  mapInputs: number;
  mapHolding: number;
  mapRegisterRead: number;
  mapRegisterWrite: number;
}

export interface ModbusRequest {
  slaveAddr: string;
  typeReq: ModbusReqType;
  firstModbusElement: number;
  nbrModbusElements: number;
  logicInverted: boolean;
  offsetVarMapped: number;
}

export interface ModbusStatus {
  running: boolean;
  currentReq: number;
  frameCount: number;
  errorCount: number;
  slavePort: number;
}